Restructure vcxagency files
